# chore(gatsby): Use options in gatsby-preset & use native loader for deps