--- a/packages/babel-preset-gatsby/src/__tests__/dependencies.js
+++ b/packages/babel-preset-gatsby/src/__tests__/dependencies.js
@@ -1,0 +1,7 @@
+const preset = require(`../dependencies`)
+
+describe(`dependencies`, () => {
+  it(`should specify proper presets and plugins when stage is %s`, () => {
+    expect(preset()).toMatchSnapshot()
+  })
+})