--- a/packages/babel-preset-gatsby/src/index.js
+++ b/packages/babel-preset-gatsby/src/index.js
@@ -29,6 +29,7 @@ const loadCachedConfig = () => {
 module.exports = function preset(_, options = {}) {
   let { targets = null } = options
 
+  // TODO(v3): Remove process.env.GATSBY_BUILD_STAGE, needs to be passed as an option
   const stage = options.stage || process.env.GATSBY_BUILD_STAGE || `test`
   const pluginBabelConfig = loadCachedConfig()
   const absoluteRuntimePath = path.dirname(