… packages/babel-preset-gatsby/src/__tests__/__snapshots__/index.js.snap
… packages/gatsby/src/redux/__tests__/__snapshots__/babelrc.js.snap
… packages/gatsby/src/utils/__tests__/__snapshots__/webpack-utils.js.snap