## Description
Did some benchmarks and tests just to make sure everything was working.
gatsbyjs.org has the biggest difference, I think because they use lots of node_modules.

Tested on moonmeister.net because it uses netlify-cms and we ran into trouble a few times before :joy:

Current master:

| site            | #1      | #2      | #3      | avg     |
|-----------------|---------|---------|---------|---------|
| starter         | 7.601s  | 7.494s  | 9.700s  | 8.265s‬  |
| moonmeister.net | 14.181s | 14.256s | 14.096s | 14.177s |
| gatsbyjs.org    | 52.747s | 46.147s | 52.273s | 50.389s |
| reactjs.org     | 21.035s | 21.995s | 18.251s | 20.427s |

This pr:

| site            | #1      | #2      | #3      | avg     |
|-----------------|---------|---------|---------|---------|
| starter         | 4.474s  | 5.798s  | 5.700s  | 5.324s‬  |
| moonmeister.net | 11.911s | 11.731s | 11.189s | 11.610s |
| gatsbyjs.org    | 34.642s | 34.384s | 34.436s | 34.487s |
| reactjs.org     | 12.583s | 15.258s | 15.248s | 14.363s |

Difference:

| site            | diff     | %    |
|-----------------|----------|------|
| starter         | -2.941s  | -35% |
| moonmeister.net | -2.567s  | -18% |
| gatsbyjs.org    | -15.902s | -31% |
| reactjs.org     | -6.064s  | -29% |

## Related Issues