@sidharthachatterjee 2019-07-19
Published in `babel-preset-gatsby@0.2.7` and `gatsby@2.13.30`