@KagamiChan 2019-07-21
seems that this change breaks my gatsby build which used to read configuration `babel.config.js`